Move old BackDoor back to teammates.test package
