+++ src/test/java/teammates/test/cases/browsertests/FeedbackConstSumOptionQuestionUiTest.java
@@ -7,7 +7,7 @@ import org.testng.annotations.Test;
 
 import teammates.common.datatransfer.questions.FeedbackConstantSumDistributePointsType;
 import teammates.common.util.Const;
-import teammates.e2e.util.BackDoor;
+import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.InstructorFeedbackEditPage;
 
 /**
