+++ src/test/java/teammates/test/cases/browsertests/InstructorFeedbackSubmitPageUiTest.java
@@ -11,7 +11,7 @@ import teammates.common.datatransfer.questions.FeedbackNumericalScaleResponseDet
 import teammates.common.util.AppUrl;
 import teammates.common.util.Const;
 import teammates.e2e.cases.e2e.BaseE2ETestCase;
-import teammates.e2e.util.BackDoor;
+import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.FeedbackSubmitPage;
 
 /**
