+++ src/test/java/teammates/test/cases/browsertests/InstructorSubmissionAdjustmentUiTest.java
@@ -16,9 +16,9 @@ import teammates.common.util.ThreadHelper;
 import teammates.common.util.retry.MaximumRetriesExceededException;
 import teammates.common.util.retry.RetryableTaskReturns;
 import teammates.e2e.cases.e2e.BaseE2ETestCase;
-import teammates.e2e.util.BackDoor;
 import teammates.e2e.util.Priority;
 import teammates.e2e.util.TestProperties;
+import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.InstructorCourseEnrollPage;
 
 /**
