+++ src/test/java/teammates/test/cases/browsertests/StudentCourseJoinConfirmationPageUiTest.java
@@ -7,8 +7,8 @@ import org.testng.annotations.Test;
 import teammates.common.util.Const;
 import teammates.common.util.ThreadHelper;
 import teammates.e2e.cases.e2e.BaseE2ETestCase;
-import teammates.e2e.util.BackDoor;
 import teammates.e2e.util.TestProperties;
+import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.AppPage;
 import teammates.test.pageobjects.StudentCourseJoinConfirmationPage;
 import teammates.test.pageobjects.StudentHomePage;
