+++ src/test/java/teammates/test/cases/browsertests/StudentFeedbackSubmitPageUiTest.java
@@ -17,8 +17,8 @@ import teammates.common.datatransfer.questions.FeedbackNumericalScaleResponseDet
 import teammates.common.util.AppUrl;
 import teammates.common.util.Const;
 import teammates.e2e.cases.e2e.BaseE2ETestCase;
-import teammates.e2e.util.BackDoor;
 import teammates.e2e.util.TestProperties;
+import teammates.test.driver.BackDoor;
 import teammates.test.pageobjects.AppPage;
 import teammates.test.pageobjects.FeedbackSessionNotVisiblePage;
 import teammates.test.pageobjects.FeedbackSubmitPage;
